requirements.txt
certifi==2019.11.28
cffi==1.13.2
fastBPE==0.1.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
pandas==1.0.0
Pillow==7.0.0
portalocker==1.5.2
protobuf==3.11.2
pycparser==2.19
python-dateutil==2.8.1
pytorch-ignite==0.3.0
pytz==2019.3
regex==2020.1.8
sacrebleu==1.4.3
six==1.14.0
tensorboardX==2.0
tqdm==4.42.0
typing==3.7.4.1
